refactor(parsers/icechunk): rename _obstore_storage to obstore_utils
Cleaner name; drops the leading underscore so the module is openly part
of the package's public surface (importable as
`virtualizarr.parsers.icechunk.obstore_utils` for anyone who wants the
translator on its own).

Also: ruff format took one line wrap-down and numpydoc nudged a test
docstring to start with a capital letter. No behavior change.
