+++ virtualizarr/parsers/icechunk/obstore_utils.py
